feat(adapter): refactor to engine, default to conventional commit types
Closes #29
